requirements.txt
pyrogram==2.0.106
tgcrypto
pymongo
python-dotenv
qrcode
pillow

